Part2 done

+++ src/main/java/facades/PersonFacade.java
@@ -1,5 +1,6 @@
 package facades;
 
+import exceptions.PersonNotFoundException;
 import dtos.PersonDTO;
 import dtos.PersonsDTO;
 import entities.Person;
@@ -17,7 +18,7 @@ import utils.EMF_Creator;
  *
  * Rename Class to a relevant name Add add relevant facade methods
  */
-public class PersonFacade {
+public class PersonFacade implements IPersonFacade {
 
     private static PersonFacade instance;
     private static EntityManagerFactory emf = EMF_Creator.createEntityManagerFactory();
@@ -42,25 +43,29 @@ public class PersonFacade {
     private EntityManager getEntityManager() {
         return emf.createEntityManager();
     }
-    
-    public PersonDTO deletePerson(Integer id) {
+
+    @Override
+    public PersonDTO deletePerson(Integer id) throws PersonNotFoundException {
         EntityManager em = getEntityManager();
 
         try {
             em.getTransaction().begin();
 
             Person person = em.find(Person.class, id);
-            em.remove(person);
-                       
-            em.getTransaction().commit();
-            
-            return new PersonDTO(person);
-                
+            if (person != null) {
+                em.remove(person);
+                em.getTransaction().commit();
+                return new PersonDTO(person);
+
+            } else {
+                throw new PersonNotFoundException(String.format("Person with id (%d) not found", id));
+            }
         } finally {
             em.close();
         }
     }
 
+    @Override
     public PersonDTO editPerson(PersonDTO p) {
 
         EntityManager em = getEntityManager();
@@ -69,30 +74,30 @@ public class PersonFacade {
             em.getTransaction().begin();
 
             Person person = em.find(Person.class, p.getId());
-            
-            if(p.getfName() != null) {
-            person.setfName(p.getfName());
+
+            if (p.getfName() != null) {
+                person.setfName(p.getfName());
             }
-            
-            if(p.getlName() != null) {
-            person.setlName(p.getlName());
+
+            if (p.getlName() != null) {
+                person.setlName(p.getlName());
             }
-            
-            if(p.getPhone() != null) {
-            person.setPhone(p.getPhone());
+
+            if (p.getPhone() != null) {
+                person.setPhone(p.getPhone());
             }
-            
+
             person.setLastEdited();
-            
+
             em.getTransaction().commit();
             return new PersonDTO(person);
-                
+
         } finally {
             em.close();
         }
     }
-    
 
+    @Override
     public PersonDTO addPerson(String fName, String lName, String phone) {
         EntityManager em = getEntityManager();
         Person person = new Person(fName, lName, phone);
@@ -106,6 +111,7 @@ public class PersonFacade {
         }
     }
 
+    @Override
     public PersonsDTO getAllPersons() {
         EntityManager em = getEntityManager();
         try {
@@ -115,11 +121,16 @@ public class PersonFacade {
         }
     }
 
-    public PersonDTO getPerson(Integer id) {
+    @Override
+    public PersonDTO getPerson(Integer id) throws PersonNotFoundException {
         EntityManager em = getEntityManager();
         try {
             Person person = em.find(Person.class, id);
-            return new PersonDTO(person);
+            if (person != null) {
+                return new PersonDTO(person);
+            } else {
+                throw new PersonNotFoundException(String.format("Person with id (%d) not found", id));
+            }
         } finally {
             em.close();
         }
